Minor change to avoid compiler warnings

+++ measCompApp/demoSrc/drv1608G.cpp
@@ -249,7 +249,7 @@ private:
   int defineWaveform(int channel);
 };
 
-#define NUM_PARAMS (&LAST_USB1608G_PARAM - &FIRST_USB1608G_PARAM + 1)
+#define NUM_PARAMS ((int)(&LAST_USB1608G_PARAM - &FIRST_USB1608G_PARAM + 1))
 
 static void pollerThreadC(void * pPvt)
 {
